some renames and small refactor in AvailableExpressionsOptimization tests
