+++ SimpleLang1.csproj
@@ -57,7 +57,7 @@
     <Compile Include="DataFlowAnalysis\ConstPropagation.cs" />
     <Compile Include="DataFlowAnalysis\ReachingDefinitionsGlobal.cs" />
     <Compile Include="DataFlowAnalysis\LiveVariableOptimization.cs" />
-    <Compile Include="DataFlowAnalysis\AvailableExpressionsApplication.cs" />
+    <Compile Include="DataFlowAnalysis\AvailableExpressionsOptimization.cs" />
     <Compile Include="LoopsInCFG\BackEdges.cs" />
     <Compile Include="LoopsInCFG\CFGRegions.cs" />
     <Compile Include="LoopsInCFG\NaturalLoop.cs" />
